Separate utils functions into their own file
